+++ src/main/java/alice/command/Command.java
@@ -1,3 +1,8 @@
+package alice.command;
+
+import alice.storage.*;
+import alice.ui.*;
+
 public abstract class Command {
     private enum Action {
         LIST,
