+++ src/main/java/alice/task/Deadline.java
@@ -1,4 +1,4 @@
-package task;
+package alice.task;
 
 import java.time.*;
 import java.time.format.*;
