+++ src/main/java/alice/task/Parser.java
@@ -1,4 +1,4 @@
-package task;
+package alice.task;
 
 import java.time.*;
 import java.time.format.*;
